Added modded marginov color picker
commit 6a0ea292c6393dcde71a91f1a26141382d917b9f
Author: Alex Cruz <mazdarider23@gmail.com>
Date:   Tue Jan 31 08:51:05 2017 +0100

    Replace colorpicker action set png with vector

    Change-Id: Ie44e1e91234706bb80f1ac56c72a68e1fe8eff07
    Signed-off-by: mydongistiny <jaysonedson@gmail.com>

commit 6da2ccafd5ed27fe0b6eca1430caa50b81f78cae
Author: Lars Greiss <kufikugel@googlemail.com>
Date:   Fri Dec 19 00:18:41 2014 -0500

    Added modded marginov color picker

    -- Adapt for Preference being part of support library

    Change-Id: Ic721fdc7f63607d972d203e1a13097e8870bce2f

BenzoEdit: Fixed force close
Change-Id: Ic574819dbd6ac31635b3f0f0684fe9febc02ea13

ColorPickerPreference: Introduce set default button when attribute is…

… set

Set a default value using the "defaultColorValue" settings namespace
attribute. The general idea is often we want set a value of say "-1"
and let framework handle it. One such scenario would be if you are
changing the color of a ImageView using ColorFilter. When framework
receives a "-1" value, framework sets the ImageView ColorFilter to null,
restoring the ImageView to original state.

Screenshot:

https://teameos.slack.com/files/bigrushdog/F09T5S155/screenshot_2015-08-31-04-02-54.png

Change-Id: I4ad63576270de344a9e7430d9e63a248d37afd9f